requirements.txt
fastapi>=0.87.0
numpy>=1.21.2
Pillow>=9.3.0
requests>=2.26.0
streamlit>=1.14.0
tensorflow>=2.9.1
uvicorn>=0.20.0
